Added Twitter api client
